+++ src/components/press-header/press-grid/press-grid-arrow.js
@@ -1,6 +1,6 @@
-import { HEADER_CLASS, NUMBER, MODE, ATTRIBUTE } from '../constants/press-grid.js';
-import { getSliceIds } from '../utils/shuffle.js';
-import { Store } from '../utils/store.js';
+import { HEADER_CLASS, NUMBER, MODE, ATTRIBUTE } from '../../../constants/press-grid.js';
+import { getSliceIds } from '../../../utils/shuffle.js';
+import { Store } from '../../../utils/store.js';
 
 class GridArrowStore extends Store {
   constructor(pressData, pressList) {
